requirements.txt
colorama
pyyaml
yagmail[all]
python-dotenv
#required package version because of a bug introduced in the version 39.0 ...
cryptography==38.0.4
